added delete for everything

+++ src/main/java/finalproject/soundcloud/controller/PlaylistController.java
@@ -7,7 +7,6 @@ import finalproject.soundcloud.model.pojos.Playlist;
 import finalproject.soundcloud.model.pojos.User;
 import finalproject.soundcloud.model.repostitories.PlaylistRepository;
 import finalproject.soundcloud.util.exceptions.DoesNotExistException;
-import finalproject.soundcloud.util.exceptions.NotLoggedException;
 import finalproject.soundcloud.util.exceptions.SoundCloudException;
 import finalproject.soundcloud.util.exceptions.UnauthorizedUserException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,20 +24,26 @@ public class PlaylistController extends SessionManagerController {
     @Autowired
     PlaylistRepository playlistRepository;
     @PostMapping(value = "users/{id}/createPlaylist")
-    public ResponseDto createPlaylist(@RequestBody PlaylistDto playlistDto, HttpSession session) throws SoundCloudException {
+    public ResponseDto createPlaylist(@RequestBody PlaylistDto playlistDto, HttpSession session,@PathVariable("id") long userId) throws SoundCloudException {
         isUserLogged(session);
         User user = (User) session.getAttribute(LOGGED);
+        if(user.getId() != userId){
+            throw new UnauthorizedUserException();
+        }
         playlistDao.createPlaylist(user,playlistDto);
         return new ResponseDto("playlist created!");
     }
-    @PostMapping(value = "users/{id}/playlists/{plId}")
-    public ResponseDto deletePlaylist( @PathVariable("plId") long playlistId,
+    @DeleteMapping(value = "users/{id}/playlists/{plId}")
+    public ResponseDto deletePlaylist( @PathVariable("plId") long playlistId,@PathVariable("id") long userId,
            HttpSession session) throws SoundCloudException {
         isUserLogged(session);
         User user = (User) session.getAttribute(LOGGED);
         Playlist playlist = playlistRepository.getById(playlistId);
         if(playlist == null){
             throw new DoesNotExistException("playlist");
+        }
+        if(user.getId() != userId){
+            throw new UnauthorizedUserException();
         }
         return playlistDao.deletePlaylist(user,playlist);
     }
@@ -56,13 +61,27 @@ public class PlaylistController extends SessionManagerController {
         return playlistDao.addSong(playlistId,songId);
     }
     @DeleteMapping(value = "users/{id}/playlists/{plId}/songs/{sId}")
-    public ResponseDto removeSongFromPlaylist(){
-
-        return new ResponseDto();
+    public ResponseDto removeSongFromPlaylist(@PathVariable("id") long userId,@PathVariable("sId") long songId,@PathVariable("plId") long playistId , HttpSession session) throws SoundCloudException {
+        isUserLogged(session);
+        User user = (User) session.getAttribute(LOGGED);
+        if(user.getId() != userId){
+            throw new UnauthorizedUserException();
+        }
+        if(playlistRepository.getById(playistId) == null){
+            throw new DoesNotExistException("playlist");
+        }
+        if(!playlistDao.removeSong(playistId,songId)){
+            return new ResponseDto("oops something went wrong! try again later");
+        }
+        return new ResponseDto("song removed from playlist!");
     }
-    @DeleteMapping(value = "users/{id}/playlists")
-    public ResponseDto deleteAllUserPlaylists(){
-        
-        return new ResponseDto();
-    }
+    /*@DeleteMapping(value = "users/{id}/playlists")
+    public ResponseDto deleteAllUserPlaylists(@PathVariable("id") long userId,HttpSession session) throws SoundCloudException {
+        isUserLogged(session);
+        User user = (User) session.getAttribute(LOGGED);
+        if(user.getId() != userId){
+            throw new UnauthorizedUserException();
+        }
+        return new ResponseDto("all playlists removed!");
+    }*/
 }
